Refactored util files to right folder
